refactor : 알맞은 테스트 위치로 이동 및 누락 테스트 추가
